Move shallowEqual to utils
